Update patch set 2
Patch Set 2: Code-Review-1

Hmm, something seems to doesn't work :/

1. Click search (button in alpha or bar in stable/beta)
-> search fades in (url gets /search hash)
2. Close the search overlay (x or click on white body)
-> search overlay fades out
Expected: /search hash should be removed
Observed: /search hash still there

And one problem with this is, that it seems, that other hashes are overwritten. E.g. as a 3rd step, try to open the editor (look at the hash and the flash of the overlay) and try to close it. In chrome it doesn't close :(

Patch-set: 2
Label: Code-Review=-1