Update patch set 2
Patch Set 2: Code-Review-1

It was a choice we made a while back to make our lives easier.

A lot of this was to do with hacks around inline styles where  we forced to 100% width that also had padding.

This will also lead to slight visual changes e.g. the main menu gets slightlier bigger.

Help me understand "potentially problematic"
We could probably scope it in such a way so that it only applies to the content/UI and doesn't interfere with VisualEditor if this is where you are seeing issues

It might also be worth using inherit [1] to allow easier overriding.
[1] http://www.paulirish.com/2012/box-sizing-border-box-ftw/

Patch-set: 2