Update patch set 2
Patch Set 2:

Rightly or wrongly content-box is the default model in user agents. That means for the most part people develop against this model so whenever we bring in external code it needs to be overridden (as you can see with CentralNotice). The number of cases where we currently combine width/height with padding/border is very low, I could only find two (we can double check with CSS lint) and didn't bother with the menu one because it's 1px. Given that the rule has been used twice in a few years it's hardly making our lives easier.

The default state should be that we don't need a hack to fix CSS because of an inherited rule, not vice-versa.

Patch-set: 2