Update patch set 2
Patch Set 2:

Seems a bit like an equivalent of global variables and functions. This border-box rule is conveniently available everywhere and MobileFrontend and saves you from having to actually figure out when you actually meant to use it. That paradigm doesn't scale. You're better off putting it where you intended to use it. Even if that's a central rule or class that's shared between a lot of MobileFrontend's layout. But not on generic selectors that you don't "own".

Patch-set: 2