Update patch set 2
Patch Set 2:

I imagine it is exactly inline styles that are least compatible with this, because those are written in isolation with no access to stylesheets and are typically written on desktop. And would probably be interpreted incorrectly on mobile due to this.

Interface components of MobileFrontend that looking different without this rule isn't a bug. Most probably any decent library written today will look different if you remove its border-box rules (OOjs UI included). It simply has to be applied in the right location instead. So plain removal of this rule (as in this commit) should probably not be merged without adding it in the proper places first (or in the same commit).

The reason I can imagine for doing this globally is to save a few bytes in download (since MF would likely never reset it much), but that'd imho be an inappropriate micro-optimisation with functional side effects.

Patch-set: 2