Update patch set 2
Patch Set 2:

I'm trying to bring as much mobile integration code into VE-core as possible to remove the amount of hacks we have in MF to get VE working, both JS and CSS hacks. Compared to desktop integration it's a mess. All of the limited UI we have in VE mobile so far (B/I/Cite) has a load of custom styling just to get it working. We have nothing like this in desktop and we shouldn't need it. As we add more functionality, this border-box problem is going to keep rearing it's head. There are *two* places where border-box changes the layout in MF - yet the price we pay for applying this rule everywhere is hundreds of potential future landmines.

Patch-set: 2