Update patch set 2
Patch Set 2: Code-Review+1

Ed, thank you for taking the time to clean up this mess.

Arguments for keeping the global box-sizing rule:
1. It's making our lives easier
2. Don't change things based on developer taste

Arguments for removing the global box-sizing rule:
1. Creates an incompatible environment for external code
2. Problem it's meant to solve can be solved in a better way
3. Breaks "the internet" by necessitating special casing

My take: I am confident that this should be merged. We should dispense with the notion that things are of value because they supposedly make our lives easier. Things should be done correctly first and foremost, and hacks should be used only when we have no other option.

The idea that this is of value because it conveniences us is problematic. First, it's dubious whether that is even true. Second, even if it does, it does so at the expense of others. Third, In most cases we are the ones suffering the externalities of this decision (see first point).

Merge it.

Patch-set: 2
Reviewer: Trevor Parscal <88@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1