Update patch set 2
Patch Set 2:

Thanks Ryan, that's a good point. I know MF is a bit more agressive with re-styling page content than just a regular skin and I don't want this patch to cause any regressions. If you could point me in the direction of which rules you think that would apply to that would be very helpful.

Patch-set: 2
Label: Verified=0