Update patch set 4
Patch Set 4:

I think as proven by the comments that this is mainly a philosophical CSS decision, and it affects not only current code but future work, as such I think it needs discussion and approval (or at least acknowledgement) of many contributors to the code base, not just us. We are now hunting for specific cases, and we may miss something, and the discussion is looking like religious arguments.

A tangible fixable bug would be, if VE wants to get rid of problems/hacks with this border-box, then change the box-sizing in reset.less

    html {
      box-sizing: border-box;
    }
    *, *:before, *:after {
      box-sizing: inherit;
    }

And then just add in VisualEditorOverlay.less

    .editor-overlay-ve {
        box-sizing: content-box;
        ...

So all visual editor DOM will be based on content-box. That would be automatically mergeable, nobody is going to oppose it, it gives MF a more resilient box-sizing reset and gives VE the advantages and basis to remove the hacks you were mentioning. From that good box-sizing we can then proceed to discuss about moving out of border-box, or encapsulate certain areas out of MF control to use content-box.

Besides what I've commented before, IMO when working in libraries like OOJS or embeddable components like VisualEditor (meant to be used as hosted code), they should protect themselves against external CSS influence as possible, thus setting a box-sizing for the library/component. box-sizing is a default exposed and commonly overridden property and you cannot count on the environment having it set as content-box, since that will make such embeddable components brittle and less portable.

Patch-set: 4