Update patch set 4
Patch Set 4:

I think enough of the major contributors are present here for this to proceed, if you feel anyone is missing please add them as a reviewer.

* { box-sizing: inherit; } is breaking change too, other libraries will assume that a box sizing rule does not apply to children.

This is not just about VE but all and any external code. We can't wrap every invocation of external CSS in a flawed hack. In the near future we may be migrating MF widgets to OOUI, in which case you'd have tens or maybe hundreds of such cases - such a solution is simply unworkable.

Patch-set: 4