Update patch set 1
Patch Set 1:

I've been looking at the EventEmitter connect function and tried to map it to the code in delegateEvents, but it doesn't work (or I don't see how). The EventEmmiter is meant for synthetic events in objects, but the events map inspired from backbone is meant for DOM events so there are some differences. Dom events map has on the keys { 'event selector': handler } but the events map that the connect function takes has { 'event': handler } since the object to listen is this, and it doesn't have a notion of DOM events and selectors. It also calls `this.on(event, method)` but we actually want to use jQuerys delegate form `this.on( event, selector, method );`.

I've cleaned the code in delegateEvents, and removed some idioms that I ported from the backbone version that don't match here. I've removed the isFunction in favor of the jquery version.

If we get to a point where we are happy I'll move this code to the View file in mantle, but let's use this for experimenting, it makes easier to review and change with this chain of commits.

About the difference between events handlers and public methods, from previous experiences most of the time the event handlers are actually methods that perform an action and should be at the view level and not mangled inside of handlers of handlers of methods. 

They can perfectly (and most of the time) be a public method, thus encapsulating specific functionality and making it more testable (no need to trigger dom events if you don't want to). In the case of actual event handlers usually prefixing them with on`Event` and organizing the code sensibly makes the trick. This also has the advantage of being able to share or invoke such handlers or methods from other methods. For example, imagine in the watchstar the new `disableClick` method can also be called from other methods or handlers, or specific actions on the view can be invoked without passing through dom events, so that is a win.

I also think that it helps and forces break down code into smaller pieces, and that is going to be a win, both on the readability and testing story. From what I've been seeing the postRender in views is usually tremendously big because of setting up events and their handlers. Finding which click does what on a big postRender (like the wikigrok ones) is kind of a nightmare.

If we get to something like this I want to slowly but surely go through views on the code that have plenty of dom event handling code and clean them up. I think it's a great idiom to have and one of the biggest wins of the backbone views.

Patch-set: 1
Label: Verified=0