Fix MW_INSTALL_PATH to default to local installation
Which is usually ../../ if executed in the local file system.

This was causing problems when running make jsduck in the pre-review script.
Locally grunt docs and MW_INSTALL_PATH=../../ make jsduck work fine.

This is the reason why 26 errors show on the docs warnings with make jsduck but
only a couple (or none with the previous patch) are shown if executed with
grunt docs.

This may be annoying if somebody works from inside vagrant?

Change-Id: I4fe8f9fb36dcf8e3bb3d1530cf6aa1eef8ddbd4b