Update patch set 4
Patch Set 4:

I'm pretty swamped with code review and other things and I promise I will look into this later when I have time and revisit the decision.

Off the top of my head I remember instances where we hit issues with a scroll bar appearing on pages due to width 100% and padding combining in horrible ways.

I'm happy to revisit this but until I'm alerted to serious bugs that this is causing it's down my priority list.

In the meantime it would help to better understand real examples where we have breakage e.g phabricator bugs.

As Joaquin suggests in the interim turning it off inside VE would be a much safer way to move this work along. If you want to submit a patch I'm more than happy to merge that.

box-sizing border-box is used all over the place as a default in many websites and in a healthy skin ecosystem for MediaWiki I can imagine many skin developers using it so I think it is in our best interests to design code that doesn't assume one over the other.

Patch-set: 4