Fix for header covering edit pencil in alpha
Moves #page-actions inside .pre-content
In tablet css, #page-actions was scoped inside .pre-content
which is strange since page-actions was not a child of pre-content
until now. Since it was overqualified, was causing override problems.
Moved it out to be consistent with the rest of the #page-action styles.
Additionally, since page-actions is no longer above pre-content in the dom,
some padding adjustments were needed on different screen sizes.

Bug: T78707
Change-Id: I8857c8f02cd93a633058817315dfc99b3bc8a25a