Update patch set 3
Patch Set 3:

I'm not sure what the box shadows are...  are you saying the drop down menu overlaps the toolbar in the desktop version too? Maybe a screenshot would help?

We should be moved towards VE being fully skinned in VisualEditor e.g. killing VisualEditorOverlay.less and using the oojs mediawiki ui theme

If this helps make that happen I'm happy to merge but I wanted to point out there is a slight styling regression (either intentional or non-intentional) in this patch.

Patch-set: 3