Update patch set 4
Patch Set 4: Code-Review-1

Fair enough. I think it's okay to go ahead with this if it simplifies our code base and makes things more readable.

I strongly recommend putting this inside the EventEmitter or View class though rather than creating a new class specifically for this (it's okay if it doesn't work and we decide to drop it but it will help with refactoring to use it).

Drop the WIP in the commit heading if you want me to review this to go in.

Patch-set: 4
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1