Update patch set 5
Patch Set 5: Code-Review-1

To be clear, this is only for version C of WikiGrok?

Regardless, this should be implemented in a separate "badges" module (with a smattering of tests (!!!)) If WikiGrokC were to raise an event when a user completed a round, then you could simply listen and respond to that event in the badges module.

Going to kick the tyres now.

Patch-set: 5