Update patch set 3
Patch Set 3: Code-Review-1

I showed this to Maryana and she wasn't quite sure about the notifications animation - when closing the notification drawer a bit of the main menu is revealed.

Can you get together with Jared and check out this with him? Once I see a +1 from Jared I'll be happy to merge.

Patch-set: 3
Label: Code-Review=-1