Hygiene: Introduce ForeignApi
Cleanup the cors cruft.
Side note: PhotoApi, WikiDataApi, NearbyApi are essentially
a SchizophrenicApi that can be either a ForeignApi or an Api
This might be a little confusing but I think we can iterate off
this now I've packaged all the similar code in the same place.

Removed useformat=mobile hack (endpoint should be mobile prefixed
if this is desired)

Tested locally for:
enwiki -> wikidata
enwiki -> commons

Change-Id: I370b301c8913c09790197a3c21fdfb4a5984b8fb