Update patch set 3
Patch Set 3:

I think we should take care of situations where the result is null. If we are not expecting a null, then nothing to worry. If a null is a possibility, then we should take care of it.

Patch-set: 3
Reviewer: Bmansurov <1963@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0