Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

This is good enough for alpha.

I'm not convinced by the jump-to-list-item behaviour when the item's index is 0 as it hides the overlay controls, which could be a little jarring for the user – try using network throttling in Chrome's DevTools.

Other than that, one super-minor inline.

Patch-set: 4
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1