Update patch set 12
Patch Set 12: Code-Review-1

Nice!

The UI feels a bit strange though.
1) I think it should be a separate overlay (#/categories/add) with heading "Add new categories to this page"
3) There should be a save button in top right that starts disabled and becomes enabled as soon as you have categories rather than hiding/showing the button when it becomes available.

Side notes:
I'm not sure about the tag button being used in this context (since it has a different effect to others) but with lack of another UI pattern to follow I guess this is fine for a first version in alpha

Patch-set: 12
Label: Code-Review=-1