Don't return a value from Schema.log
EventLogging's mw.eventLog.logEvent() returns a deferred that is meant to be
resolved if the logging call wass successful and rejected otherwise. I'd like
to do away with that return value for two reasons: (a) when we use
navigator.sendBeacon, there is no way to find out the ultimate status of the
request; (b) there is nothing useful the calling code can do with the result,
anyway.

MobileFrontend doesn't appear to do anything with the deferred return value, so
it can be removed quite easily.

Change-Id: Ia383d091e5a55c32287c650f9bac44d7a21f6fc6