Update patch set 2
Patch Set 2:

I'm not sure what a 'view' is in mobile, but I think ultimately we should be aiming to have a solution where box-sizing is only ever set in a targeted manner (i.e. not with * child selectors). If this experiment helps us work out what to do to get closer to that then I'm all for it.

Patch-set: 2
Label: Code-Review=0