Update patch set 1
Patch Set 1:

Hm.. was this meant to be in a directory of sorts? Having a jsduckCustomTags.rb file in the repo root is a bit odd. Since it's not an configuration file, entry point or metadata file, it should probably be in a subdirectory.

As for sharing the code, it's relatively minimal (though between JSDuck 4 and 5, the plugin requirements became more verbose so the code got longer). I'm not sure it's worth the overhead trying to centralise this. It's basically just configuration. Which we actually want to keep local instead of central to avoid master-regression breakages when upgrading.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0