Update patch set 1
Patch Set 1: Code-Review-1

Since in alpha we show the WIkiGrok menu item in the sidebar, we need this module so that we can use it to load the next page. I think this should be turned off only when the sidebar menu item is disabled.

Patch-set: 1
Reviewer: Bmansurov <1963@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1