Group categories to content based and administrational ones
The important categories for the user are the content based categories.

These categories are grouped as "Content based" and are the standard view
for the user.

Hidden categories (categories supposed to be for organisation only) are
grouped as "Organizational" and accessable with a tab menu. If there are
no content based categories, show the hidden as default.

Bug: T86496
Change-Id: Ice6286bf20b1e41b5c196b59be95cfd79d5e69e2