Update patch set 14
Patch Set 14:

Okay uh, so I changed my mind about how we're going to do the LookupInputWidget change, and now the plan includes renaming it and *then* fixing the weirdnesses. Patch is pending: https://gerrit.wikimedia.org/r/#/c/185254/ we'll probably release it with a new OOUI version today and you'll be able to start depending on it. I'll update this patch for you :)

Patch-set: 14
Label: Verified=0