Update patch set 12
Patch Set 12: Code-Review-1

I'd suggest we make this beta only to start with (move methods to ReferenceDrawerBeta) and play with it.

I'm seeing weird behaviour on Chrome when I scroll to the end the entire window scrolls which to me is unexpected behaviour.

I also worry about how this impacts UX this essentially turns off scrolling when the reference drawer is open so we should be more careful about making this change - at least let Maryana and Jared etc test it out in beta before pushing to all our users.

Patch-set: 12
Label: Code-Review=-1