Update patch set 12
Patch Set 12:

If you're seeing the window scroll when scrolling to the end of the drawer, this is not intended.  I tested very thoroughly in chrome and this was not the case for me.  Are you sure no cache issues?  Also this was tested and working in both chrome and IOS8 simulator.  I will move to beta and re-test before pushing again. Preventing the window from scrolling while the drawer is open is part of the acceptance criteria.  IMO It would be very disappointing if this is undesired as much thought and time has gone into this.

Patch-set: 12
Label: Verified=0