Add css classes explicitly
It's much more confusing and leaves place for regressions, if classes are
added to the output html in template's logic or implitly. Add css classes to
templates in the JS code using the templates makes it much clearer what will be
passed to the template and outputted.

Follow up: I84cc5080c36e259074f75fa42f97f2f0fc6ebbf2

Bug: T87252
Change-Id: I91c33796375c18723ea2eda1bb0bede1d037e09f