Update patch set 3
Patch Set 3: Code-Review-1

Purely because the name is confusing. We already have a tutorials folder and ContentOverlay PageActionOverlay and uploads/UploadTutorial. It's not clear how this fits in with those and looking at the card it seems completely inconsistent... I would say this isn't a tutorial but looks more like the wikigrok InfoOverlay. It just has information not steps to doing something
https://trello.com/c/uZ9geKCR/15-3-tell-people-about-what-contribute-is-when-they-click-on-it-in-left-nav 

Also I was unable to test this as I can't remember how to setup this. Could you document on wiki and link to it in commit message?

PS. Why are we not doing i18n yet? In this case it would be super cheap to do and would get us a higher amount of users (more a moan at Kaldari and Maryana here)

Patch-set: 3
Label: Code-Review=-1