Update patch set 4
Patch Set 4:

I'm arguing the architecture is ready for this right now. There is nothing in this patch that screams difficult to translate.

Even if not exposed you are creating more work for yourselves on the long term by not doing this translation now whilst it is easy. For example the infobox stuff is fully translated across all our projects even though it is experimental.

Anyway.. i'm not blocking on that point. I just wanted to feed that back. It's up to Maryana/Kaldari to decide if that's the case. I just worry that no translation is becoming the norm. I understand Amire is talking to Kaldari about i18n anyhow.

Patch-set: 4
Label: Verified=0