Update patch set 10
Patch Set 10: Code-Review+1

The buttons are fixed, functionality works, there is a slight quirk where uploading an image will reload the page and that may reload the page with the tutorial url which  may show again, but that is a different problem. I'd be fine with creating a bug for that, and merging this change.

What do you think jdlrobson

Patch-set: 10