Update patch set 8
Patch Set 8:

agreed with phuedx. that's how it is going to be. i don't think it matters for the mvp that if a user is going to be editing a list at the same time in two different places there will be race conditions. also this will be fixed when using a DB

Patch-set: 8
Reviewer: Jhernandez <2027@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0