Turning off page-impression logging for now
Not needed for any immediate tests and we can't run WikiGrok at 100%
with page-impression logging turned on. Leaving in the codebase with
instructions since we will need this for the position tests.

Also moving out of the 'show' function to make sure we catch the page
impression even if WikiGrok doesn't display for some reason (per
discussion with Analytics).

Change-Id: If8db28974470dd50b9adb958d6a8ae3068132f31