Update patch set 7
Patch Set 7: Code-Review-1

(11 comments)

I've made a through code review, this is a hard patch (lots of small things and corner cases) and I don't think our tests are extensive enough to pick up if we fuck it up. Left a bunch of comments there.

Everytime a view gets events map, we have to check if the parent has it, and extend instead of defining it, and check other views which extend the current one, and if they use events map then extend the parent one that we just added.

I'm going to write something down and send an email to mobile-l later, I don't think we have discussed this gotchas before this patch.

Patch-set: 7
Label: Code-Review=-1