Update patch set 3
Patch Set 3: Code-Review-1

A couple issues with the spinner...
In beta mode, but not alpha, the spinner that is displayed between page loads is left aligned instead of centered on the page.
The spinner in the WikiGrok dialog should be relocated from the middle of the dialog to the 'Next' button.
The spinner should replace the button text when displayed (instead of being added to the side of it).

Patch-set: 3
Label: Code-Review=-1