Update patch set 1
Patch Set 1:

correct. If you never clicked on the watchlist (so you're a new user e.g.), you will be redirected to the default watchlist. For all future views (and links in the main menu), the target view will be read from the users option (if the user click on the a-z view, the link in the main menu will take him to the a-z view, same with the feed view). The default view is just for the first visit of the watchlist and there should be no difference, if the first view is a direct navigation to Special:Watchlist (e.g. through the url) or the sidebar :)

Patch-set: 1