Update patch set 2
Patch Set 2:

Yes, I think the buttons will looks similar once we completely switch to using oojs-ui. Also, I was hoping to put other kinds of buttons in the Button.js so that we don't create multiple files for multiple ui elements. Or is it OK to create multiple button files?

Patch-set: 2