Update patch set 2
Patch Set 2: Code-Review-1

> Or is it OK to create multiple button files
The actual file organisation is like phuedx said (whether that is useful or not is another question :P), but i think we should be consistent. It would be terrible to have many different file organisations and i have to look again all over the time where which class is defined. Maybe it would be good in a seperate directory javascripts/buttons/ e.g..

I think it would be better, if we change all visible buttons (at least for stable mode) in one change, to hold this widely used mode consistent and don't confuse people with different button styles in one document :)

Patch-set: 2
Label: Code-Review=-1