Update patch set 1
Patch Set 1:

> Also I think this will require a cache flush or a JS workaround to inject the button on pages where it is not present.
Just, if we don't want to wait. Actually the talk button doesn't appear on any page except beta and up, so there wouldn't be a problem, if we just wait? :/

> I do worry this is a premature move.
I understand what you mean, but it's (from my point of view) too long, that MF doesn't has the ability to easily show talk pages :( Do you think it would be better to limit it to logged in users (they normally understand how talk pages work or it doesn't really matter, if they see the desktop or mobile version?).

I'm not really sure, what blocks us from promoting the talk overlay feature to stable, iirc "just" the answer/contribute functionality, right? I'm not sure, if we ever really focus on this, talk pages are a difficult topic, if we want to build a really good overlay, it's the false backend. Maybe Flow is the better alternative?

What i want to say: A link to a talk page in wikitext is better as no talk button. But that's my point of view :)

Patch-set: 1