Update patch set 1
Patch Set 1: Code-Review-1

So I talked to Maryana about this. The reason we never promoted this was because we were not clear how Flow fitted into mobile and the talk page experience and editors who use talk could get it by opting into beta or user js.

If a new user clicks on the talk button and doesn't find it easy to understand they may never use it again. I think if we do promote it thus we should do it with the talk overlay which presents the content and workflow much clearer. Can you rewrite the patch to also promote the talk overlay and take into account cache issues? The talk overlay will also need some @smoke browser tests since it is being pushed to stable (any feature pushed to stable should come with browser tests / unit tests so we can keep it stable).

Patch-set: 1
Label: Code-Review=-1