Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Please do not use Special:MobileMenu
It's only supposed to be used for non-JavaScript users - there is a flash of unstyled content that shows this.

Talking to Maryana it sounds like you want to build a new special page, as the search bar is supposed to show at the top.

Also this should probably be loaded via JavaScript as this introduces a delay in opening the menu, but not being on your team and not knowing what card you are working on I'm not of the end goal here but this current experience feels very wrong..

Patch-set: 1
Label: Code-Review=-1