Update patch set 1
Patch Set 1:

Yes, you're right. It depends on [1]. I forgot to add a WIP. I felt the same about the delay, but since it's in beta, we decided to go with it. Thanks for the review. Also, these and other features are on our trello board [2]. You're welcome to see what card requirements are.

[1] https://trello.com/c/Isf8stWH/26-5-new-mobile-menu-page
[2] https://trello.com/b/6c9m1gEh/wikigrok-current-sprint-41-oldboy

Patch-set: 1
Label: Verified=0