Update patch set 4
Patch Set 4: Code-Review-1

> Ideally the banner image should be available by default without this API request

It sounds false to use the API for an action that doesn't require some user input or any other client side information. The Api can be used in PHP, too (or at least: it should be), so we should think about to deliver the banner image directly in the Html output, instead of making an Api request.

Patch-set: 4