Update patch set 5
Patch Set 5:

I don't think we should hobble Special:UserProfile like this. Not every wiki has the issues or attitudes of Wikimedia wikis. Instead of ripping out part of a Special page that is already kind of limited in terms of functionality, let's put Special:UserProfile in its own MediaWiki extension and figure out what to do with it later. That is, <https://phabricator.wikimedia.org/T85929>. There's already near universal agreement that Special:UserProfile shouldn't be living in this extension anyway, so we'll solve multiple problems at once.

This is a -2 from me for this approach. I would be happy to support a changeset that creates a new separate extension.

Patch-set: 5
Reviewer: MZMcBride <531@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0