Remove Special:UserProfile functionality
In order to not break links in cached pages,
Special:UserProfile now redirects to the user's
user page if they are a registered user, or Special:Contributions if they are
a logged out user, which is the same logic as Linker::userLink().

Changes:
* Promote beta feature
* Remove user profile special page and its browser tests

Future:
* When merging this patch please create a card to remove the Special:UserProfile
page in 30 days when redirects are no longer a problem!

Bug: T120685
Change-Id: Ie5fea5f51dcf1816628155bb401a200516d66a60