Update patch set 3
Patch Set 3: Code-Review-2

Not quite ready for this yet. 

Right now, various links on the mobile site (the username in the last modified strapline, the diff view in the mobile watchlist and history page) point to the mobile user profile. If we remove the link, we'll need to replace it with something else, and it's not immediately clear what that "something else" should be on mobile to provide value to users for all the use-cases they might tap on that link for. Contribution pages and user pages make sense on desktop, but they're not very mobile-friendly in the best of cases and completely broken-looking in the worst.

Patch-set: 3