Update patch set 3
Patch Set 3: Code-Review-1

I think this is a bit of yet another overreaction (i mean this feature has been live for 2 years why the hoo ha now? Also we can't just remove stuff on whims we should be more days driven otherwise our projects are going to end up as the site that never changes and nobody edits.

From my view at minimum this should be ported to another extension (possibly renamed user summary/dashboard as this is essentially what it is doing (on hindsight the name profile was a bad choice).

All user links should point to the contributions page when the extension is not installed since that is mobile optimised (and has a link to a user page). As maryana points out user pages suck on mobile and since the experience is inconsistent it can confuse people (we have ux research now if you want to poke them). If you really want to invest your energises into a battle of removing it then you can do it when it's in that form.

The only issue i see with the existing feature is the patch dan Garry started which is basically done. That seems like the easiest thing to do right now.

Patch-set: 3