Resolve router deferred on back
IE has a problem with correctly emitting 'popstate' after history.back().
See https://connect.microsoft.com/IE/feedback/details/793618/history-back-popstate-not-working-as-expected-in-webview-control
This patch resolves the $.Deferred in router.back() if onpopstate
doesn't fire for some reason (i.e. bug in browser or browsers that don't
support popstate). The patch is not reinventing anything, it's using the
same setTimeout function as before but extending its scope to cater for
buggy implementations of 'onpopstate'.

Bug: T91421
Change-Id: Ib1bd861f878c1538a28f3bf25250c52f4449911e