Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

Thanks for this Florian.

I worry the warning might present a ux problem and would strongly urge this is done in A/B test form so you can compare which is more effective. It obstructs the textarea and continue button, in addition to looking very scary :-) and may result in __less__ edits.

In addition to this is highly likely to dramatically reduce sign ups / logins and maybe even edits (some people may not want to give their IP and not realise logging in / signing up protects them and is available).

The UI is also very inconsistent from the rest of the mobile UI.

I also wonder if legally we even need it given there is a warning on the edit summary screen.

Italian Wikipedia may be willing to take this hit in the name of science/learning but I'm not sure about other wikis, and given the RFC pushing to enable this everywhere I think we need to be much more cautious. It should at least be behind a flag.

Please also see Jaredzimmerman-WMF  comment on the bug in Phabricator.
Editing is our bread and butter and we really must be more careful whilst fiddling with it. Let's discuss it more there before trying to push through  changes like this.

Personally I see the CTA as a good filter. It reminds people that they are editing anonymously and to be careful. I'm not convinced it is blocking edits and if it is it is probably blocking the right sort of edits.

Patch-set: 5
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1