Update patch set 3
Patch Set 3:

Jon: I don't understand your comment. In your final paragraph, you say the only issue you see is the inclusion of images, but in the preceding paragraphs you agree that the name "UserProfile" is less than ideal.

I thought there was common agreement that MobileFrontend is the wrong extension for Special:UserProfile. What harm is there in moving this code out of an inappropriate place? If there's interest in (another) separate extension, the Git history is available to all.

Patch-set: 3