Update patch set 1
Patch Set 1: Code-Review-1

(7 comments)

This is a good start. My comments aren't about the approach, but about the organisation of the code.

Also, as regards not knowing the dimensions of the device upfront: an interesting follow-on patch would be integrating an existing device detection library to try and make an educated guess (for iDevices etc.)

Patch-set: 1
Label: Code-Review=-1