QA: recent versions of Chrome don't wait for event listeners
The only workaround we've found is a sleep(). I am
investigating filing an upstream bug for this.

This patch should address the failures seen in
https://integration.wikimedia.org/ci/job/browsertests-MobileFrontend-SmokeTests-linux-chrome-sauce/49/
However, I have only been able to reproduce the issue locally for
notifications.feature. It may also be necessary to add this for
language.feature as well, but I haven't done that yet.

Change-Id: I0d1aeb5fd1966a9e22890455a1607ae1339d83fe