Update patch set 2
Patch Set 2: Code-Review-1

I still don't get this... if this is for browsers which do not fire popstate we should feature / browser detect against those browsers - not do this for everything.

This is a core piece of code and I don't want to risk breaking the experience of other browsers for IE11.

Do we know why onpopstate doesn't fire?

Is this based on this:
http://stackoverflow.com/questions/22862749/oppopstate-not-trigger-in-firefox-and-ie11

or some other blog?
More information please!

Patch-set: 2
Label: Code-Review=-1