Update patch set 2
Patch Set 2:

OK, let me first clarify that I'm not radically changing the existing behavior and there is no need to panic. All existing tests are passing and I've even written new ones. That said, I'm extending the existing behavior because it was buggy, it relied on the fact that the popstate would fire when history.back() is run, which is obviously not the case. So I made it so that if onpopstate doesn't fire, then still resolve the deferred object. Your suggested solution of feature detection does not apply here because even though 'onpopstate' exists in later versions of IE, its behavior is different from the one in other browsers.

I was able to find some info about why it doesn't fire on IE.
"The window.onpopstate event handler notifies your app when the user navigates between two travel entries in the history, where at least one was created with either history.pushState() or history.replaceState(). This occurs through UI interaction, such as clicking the back or forward button, clicking Back from a context menu, or pressing the backspace key. It also can occur by pressing alt+left, alt+right, or the backspace key, or by calling history.back, history.forward, or history.go."

The important piece here is "This occurs through UI interaction, such as clicking the back or forward button, clicking Back from a context menu, or pressing the backspace key. It also can occur by pressing alt+left, alt+right, or the backspace key, or by calling history.back, history.forward, or history.go." which is not the case with our overlays. We just override the hash which opens the overlay. More info can be found here: [1]. 

[1] https://msdn.microsoft.com/en-us/library/ie/hh920758%28v=vs.85%29.aspx

Patch-set: 2
Label: Verified=0