Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

See inline for a couple of minors.

Regarding the approach: invoking the mobileview API multiple times per page is a scalable solution to this problem. Releasing this to stable might start a fire, so I'd recommend making the relevant to SkinMinervaAlpha. Moving from beta and eventually to stable is going to require extraction of the code that sizes images from the mobileview API into its own objects, which can be shared between the two, so that we're not invoking /who knows what else/ to do one thing.

Patch-set: 4
Label: Code-Review=-1